requirements.txt
beautifulsoup4==4.10.0
opencv_python==4.5.3.56
Pillow==8.4.0
selenium==3.141.0
webdriver_manager==3.5.2
